requirements.txt
anyio==4.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cfenv==0.5.3
cffi==1.16.0
click==8.1.7
cryptography==41.0.5
deprecation==2.1.0
exceptiongroup==1.2.0
Flask==3.0.0
furl==2.1.3
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
orderedmultidict==1.0.1
packaging==23.2
pycparser==2.21
PyJWT==2.8.0
sap-xssec==4.1.0
six==1.16.0
sniffio==1.3.0
urllib3==2.1.0
Werkzeug==3.0.1
